+++ src/ClassicUO.Client/Network/PacketHandlers/Helpers/SharedStore.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Linq;
+using ClassicUO.Game;
 
 namespace ClassicUO.Network.PacketHandlers.Helpers;
 
@@ -24,5 +25,32 @@ internal static class SharedStore
             return;
 
         _customHouseRequests.Add(serial);
+    }
+
+    public static void SendMegaCliLocRequests(World world)
+    {
+        if (world.ClientFeatures.TooltipsEnabled && _cliLocRequests.Count != 0)
+        {
+            if (Client.Game.UO.Version >= Utility.ClientVersion.CV_5090)
+            {
+                if (_cliLocRequests.Count != 0)
+                    AsyncNetClient.Socket.Send_MegaClilocRequest(_cliLocRequests);
+            }
+            else
+            {
+                foreach (uint serial in _cliLocRequests)
+                    AsyncNetClient.Socket.Send_MegaClilocRequest_Old(serial);
+
+                _cliLocRequests.Clear();
+            }
+        }
+
+        if (_customHouseRequests.Count > 0)
+        {
+            for (int i = 0; i < _customHouseRequests.Count; ++i)
+                AsyncNetClient.Socket.Send_CustomHouseDataRequest(_customHouseRequests[i]);
+
+            _customHouseRequests.Clear();
+        }
     }
 }
